money set int

+++ FrontEnd/lib/model/user/user_model.dart
@@ -1,7 +1,7 @@
 class UserModel {
   String username;
   String phone;
-  double money;
+  int money;
   String role;
 
   UserModel({
